hints: fix deleting deleting hint files on drain
The send_one_file function works in two modes:

- When the endpoint manager is not draining, it tries to send all hints
  from the file. If some hints fail to be sent, it reports an error. The
  file is deleted if everything was sent successfully.
- When the endpoint manager is draining, it also tries to send all hints
  but it ignores any errors. It always deletes the file at the end.

Another function, send_hints_maybe, calls send_one_file in a loop and
stops after all segments were removed or an error occured. Recently, the
loop condition was modified so that it stops if can_send() returns
false, which could happen if the destination replica is reported in
gossip as DOWN. As send_hints_maybe is used in draining, this could
cause the loop to stop too early and leave some hint files on the disk.
This, in turn, would cause the cleanup to report an error as it tries to
remove a directory which is not completely empty.

To fix this, can_send() was removed from the loop condition. Instead, a
check is added in send_one_file which causes reading from a hints file
to be skipped in case can_send() is false. The logic related to file
removal should work correctly now.